Move file to the correct module
