+++ detekt-psi-utils/api/detekt-psi-utils.api
@@ -13,6 +13,16 @@ public final class io/github/detekt/psi/FilePath$Companion {
 	public final fun fromRelative (Ljava/nio/file/Path;Ljava/nio/file/Path;)Lio/github/detekt/psi/FilePath;
 }
 
+public abstract class io/github/detekt/psi/FunctionMatcher {
+	public static final field Companion Lio/github/detekt/psi/FunctionMatcher$Companion;
+	public abstract fun match (Lorg/jetbrains/kotlin/descriptors/CallableDescriptor;)Z
+	public abstract fun match (Lorg/jetbrains/kotlin/psi/KtNamedFunction;Lorg/jetbrains/kotlin/resolve/BindingContext;)Z
+}
+
+public final class io/github/detekt/psi/FunctionMatcher$Companion {
+	public final fun fromFunctionSignature (Ljava/lang/String;)Lio/github/detekt/psi/FunctionMatcher;
+}
+
 public final class io/github/detekt/psi/KeysKt {
 	public static final fun getBasePath (Lorg/jetbrains/kotlin/com/intellij/psi/PsiFile;)Ljava/nio/file/Path;
 	public static final fun getLineSeparator (Lorg/jetbrains/kotlin/com/intellij/psi/PsiFile;)Ljava/lang/String;
