+++ detekt-rules-complexity/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/complexity/NestedScopeFunctions.kt
@@ -1,6 +1,6 @@
 package io.gitlab.arturbosch.detekt.rules.complexity
 
-import io.github.detekt.tooling.api.FunctionMatcher
+import io.github.detekt.psi.FunctionMatcher
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.Configuration
 import io.gitlab.arturbosch.detekt.api.DetektVisitor
