+++ detekt-rules-style/build.gradle.kts
@@ -5,7 +5,6 @@ plugins {
 dependencies {
     compileOnly(projects.detektApi)
     compileOnly(projects.detektMetrics)
-    compileOnly(projects.detektTooling)
     testImplementation(projects.detektMetrics)
     testImplementation(projects.detektTest)
     testImplementation(libs.assertj)
