+++ detekt-tooling/api/detekt-tooling.api
@@ -71,16 +71,6 @@ public final class io/github/detekt/tooling/api/DetektProvider$DefaultImpls {
 	public static fun getPriority (Lio/github/detekt/tooling/api/DetektProvider;)I
 }
 
-public abstract class io/github/detekt/tooling/api/FunctionMatcher {
-	public static final field Companion Lio/github/detekt/tooling/api/FunctionMatcher$Companion;
-	public abstract fun match (Lorg/jetbrains/kotlin/descriptors/CallableDescriptor;)Z
-	public abstract fun match (Lorg/jetbrains/kotlin/psi/KtNamedFunction;Lorg/jetbrains/kotlin/resolve/BindingContext;)Z
-}
-
-public final class io/github/detekt/tooling/api/FunctionMatcher$Companion {
-	public final fun fromFunctionSignature (Ljava/lang/String;)Lio/github/detekt/tooling/api/FunctionMatcher;
-}
-
 public final class io/github/detekt/tooling/api/InvalidConfig : io/github/detekt/tooling/api/DetektError {
 	public fun <init> (Ljava/lang/String;)V
 }
